requirements.txt
python-dotenv==1.0.1
flask==3.0
requests==2.31.0
urllib3==2.2.3
